// src/project/ws/app/src/lib/routes/search/models/search.model.ts
export interface ISearchdateModified {
  from: string
  to: string
}
export type TSocialPostKind =
  | 'Blog'
  | 'Leadership-Blog'
  | 'Poll'
  | 'Query'
  | 'Reply'
  | 'Comment'
  | 'Unknown'
export interface ISearchSocialSearchPartialRequest {
  userId: string
  query: string
  pageNo: number
  pageSize: number
  sessionId: number
  postKind: TSocialPostKind
  filters?: {
    hashTags?: string[]
    tags?: string[]
    hasAcceptedAnswer?: boolean[]
    sourceName?: string[]
    sourceId?: string[]
    threadContributors?: string[]
    postCreator?: string[]
    dtLastModified?: ISearchdateModified[]
  }
  sort?: { [key: string]: string }[]
  locale?: string[]
}
export interface ISocialSearchRequest extends ISearchSocialSearchPartialRequest {
  org: string | null
  rootOrg: string | null
}
export interface IPostCreator {
  postCreatorId: string
  name: string
  emailId: string
}

export interface IThreadContributor {
  threadContributorId: string
  name: string
  emailId: string
}
export interface IPostActivity {
  activityData: {
    like: number
    upVote: number
    downVote: number
    flag: number
  }
  userActivity: {
    like: boolean
    upVote: boolean
    downVote: boolean
    flag: boolean
  }
}

export interface IWsSocialSearchResultData {
  abstract: string
  accessPaths: string[]
  activity: IPostActivity
  activityEndDate: string | null
  attachments: []
  body: string
  dtCreated: string
  dtLastModified: string
  dtPublished: string
  hasAcceptedAnswer: boolean
  highlight: {
    tags?: string[]
    body?: string[]
    title?: string[]
  }
  hashTags: string[]
  id: string
  likes: string
  options: []
  org: string
  postCreator: IPostCreator
  postKind: TSocialPostKind
  reply: []
  replyCount: number
  rootOrg: string
  source: {
    id: string
    name: string
  }
  status: string
  tags: string[]
  threadContributors: IThreadContributor[]
  thumbnail: string
  title: string
  upVoteCount: number
}
export interface ISocialFilterContent {
  displayName: string
  type: string
  count: number
}
export interface ISocialFilters {
  displayName: string
  type: string
  content: ISocialFilterContent[]
}
export interface ISocialSearchResult {
  total: number
  result: IWsSocialSearchResultData[]
  filters: ISocialFilters[]
}
// ...
export interface ISearchAutoComplete {
  _source: {
    searchTerm: string
  }
}